backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.1
tenacity==8.2.3
pydantic-settings==2.2.1
cryptography==42.0.5
watchdog==4.0.0
alembic==1.13.1
